% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneAnatomyParty.R
\name{GeneAnatomyParty}
\alias{GeneAnatomyParty}
\title{Title}
\usage{
GeneAnatomyParty(listOfSpecies, feature)
}
\arguments{
\item{listOfSpecies}{the Vector containing Species you want for a feature}

\item{feature}{The feature that you want to see}
}
\value{
Does not return only shows the plots
}
\description{
Title
}
\examples{
\dontrun{
GeneAnatomyParty("Danio_rerio","Zebrafish","pectoral fin")
}
}
