% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenesFromAnatomy.R
\name{GenesFromAnatomy}
\alias{GenesFromAnatomy}
\title{Create a radar graph of anatomical features and their genes}
\usage{
GenesFromAnatomy(speciesBgee, speciesCommon, feature)
}
\arguments{
\item{speciesBgee}{The Scientific name of a Species found in BgeeDB}

\item{speciesCommon}{The common name of the same species in Ensemble}

\item{feature}{The feature, an adjective that describes what we want to find in the
anatomical features for and how many genes we have for them}
}
\value{
a radar graph that shows the terms and how many genes we find per term
}
\description{
This function takes a species from BgeeDB and creates a Radar graph that
displays the number of genes along with the anatomica features
}
\examples{
\dontrun{
AnatomyGeneRadarGraph <-GenesFromAnatomy("Danio_rerio","Zebrafish","pectoral fin")
}
}
