% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNamesVec.R
\name{GetNamesVec}
\alias{GetNamesVec}
\title{GetNamesVec}
\usage{
GetNamesVec(name)
}
\arguments{
\item{name}{The feature that you want to see}
}
\value{
Does not return only shows the plots
}
\description{
GetNamesVec
}
\examples{
\dontrun{
NameVec<-GetNamesVec("Danio_rerio")



}
}
