% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRelevantNameVector.R
\name{GetRelevantNameVector}
\alias{GetRelevantNameVector}
\title{Converts a name into into the correct format}
\source{
\url{http://bioconductor.org/packages/release/bioc/vignettes/BgeeDB/inst/doc/BgeeDB_Manual.html}
}
\usage{
GetRelevantNameVector(name)
}
\arguments{
\item{name}{the name a user has inputed}
}
\value{
the name the functions in this package use to refer to that animal
}
\description{
This function takes a name and matches it to a species
}
\examples{
\dontrun{
Species <-GenesFromAnatomy("Danio_rerio")
}
}
