% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetUberonIds.R
\name{GetUberonIds}
\alias{GetUberonIds}
\title{Title}
\source{
\url{http://bioconductor.org/packages/release/bioc/vignettes/BgeeDB/inst/doc/BgeeDB_Manual.html}
}
\usage{
GetUberonIds(speciesBgee, speciesCommon, feature)
}
\arguments{
\item{speciesBgee}{The Scientific name of a Species found in BgeeDB}

\item{speciesCommon}{The common name of the same species in Ensemble}

\item{feature}{The feature, an adjective that describes what we want
to find in the anatomical features for and how many
genes we have for them}
}
\value{
A Dataframe with two columns one with the UBERONIds and the other
with the Anatomical feature they correspond to
}
\description{
Title
}
\examples{
\dontrun{
IdDataFrame<-GetUberonIds("Danio_rerio","Zebrafish","pectoral fin")
}

}
