% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListPhenotypesDescriptions.R
\name{ListPhenotypesDescriptions}
\alias{ListPhenotypesDescriptions}
\title{This function list the phenotype descriptions that ensembl uses that can be used as feature paramters}
\usage{
ListPhenotypesDescriptions()
}
\value{
list of Phenotype Descriptions
}
\description{
This function list the phenotype descriptions that ensembl uses that can be used as feature paramters
}
\examples{
List<-ListPhenotypesDescriptions()
}
