% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListSpecies.R
\name{ListSpecies}
\alias{ListSpecies}
\title{Show all The animals available from BgeeDB}
\usage{
ListSpecies()
}
\value{
A list of species that currently work with this package, they have phenotype descriptions
in ensemble but also are in the BgeeDB database of species both are required for this to work
}
\description{
Show all The animals available from BgeeDB
}
\examples{
List<-ListSpecies()
}
