% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phyenotypeDescriptions}
\alias{phyenotypeDescriptions}
\title{Phenotype Descriptions from Ensemble.}
\format{
A list with 43366 entries
\describe{
\item{description}{description, of a Phenotype}
}
}
\usage{
phyenotypeDescriptions
}
\description{
A vector containing the prices and other attributes of almost 54,000
diamonds.
}
\details{
#' @source \url{https://uswest.ensembl.org/index.html}
}
\examples{
\dontrun{
 phyenotypeDescriptions
}

}
\keyword{datasets}
