% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runZooGVT.R
\name{runZooGVT}
\alias{runZooGVT}
\title{launch a Shiny App for ZooGVT}
\usage{
runZooGVT()
}
\value{
This function just to run the Shiny page for this app and has no return values
}
\description{
A function that launches the Shiny app for this package
The code has been placed in \code{./inst/shiny-scripts} in a file called app.R
}
\examples{
\dontrun{

ZooGVT::runZooGVT()
}
}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
Anjali Silva (2020). TestingPackage \href{https://github.com/anjalisilva/TestingPackage/}{Link}
}
