% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Species}
\alias{Species}
\title{Phenotype Descriptions from Ensemble.}
\format{
A list of the names of Species that works with this package
\describe{
\item{Common name}{common name of the Species}
\item{Sceienticifc name}{Scientific name that will be the BgeeDB parmater most of the time}
}
}
\usage{
Species
}
\description{
A vector containing the prices and other attributes of almost 54,000
diamonds.
}
\details{
#'
}
\examples{
\dontrun{
 Species
}

}
\keyword{datasets}
